+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add a trace visualization script `tracevis.py`
 - Add `config` flag to set specific MemPool flavor, either `minpool` or `mempool`
 - Add bypass channels through the groups for the northeast intergroup connection
+- Add capability to quickly write a value via a CSR
 
 ### Fixed
 - Avoid the elaboration of SVA assertions on the `reorder_buffer` module
@@ -38,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Register wake-up signals and use `wfi` for barriers
 - Bump the dependencies to the latest version (`common_cells`, `register_interface`, `axi`, `tech_cells_generic`)
 - Use the latest version of Modelsim by default
+- Consistently print Verilator's simulation time in decimal
 
 ## 0.4.0 - 2021-07-01